LayoutTests/accessibility/custom-elements/menuitem-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("menu-1").selectedChildrenCount is 1
PASS platformValueForW3CName(accessibilityController.accessibleElementById("menu-1").selectedChildAtIndex(0)) is "item 2"
PASS accessibilityController.accessibleElementById("menu-1").selectedChildAtIndex(0).isSelected is true
PASS accessibilityController.accessibleElementById("menu-2").selectedChildrenCount is 1
PASS platformValueForW3CName(accessibilityController.accessibleElementById("menu-2").selectedChildAtIndex(0)) is "item 3"
PASS accessibilityController.accessibleElementById("menu-2").selectedChildAtIndex(0).isSelected is true
PASS accessibilityController.accessibleElementById("item-1").isEnabled is false
PASS accessibilityController.accessibleElementById("item-1").isExpanded is true
PASS accessibilityController.accessibleElementById("item-1").hasPopup is true
PASS accessibilityController.accessibleElementById("item-2").isEnabled is true
PASS accessibilityController.accessibleElementById("item-2").isExpanded is false
PASS accessibilityController.accessibleElementById("item-2").hasPopup is false
PASS successfullyParsed is true

TEST COMPLETE
